refactor: separar backend em 4 microservices deployaveis (Fly.io)
- shared/config.py: settings comuns aos 4 servicos
- gateway/auth/mobilidade/colaboracao: main.py, Dockerfile, fly.toml e requirements.txt proprios
- remove monolito main.py/Dockerfile/fly.toml da raiz de src/backend
- models/ e alembic/ seguem centralizados (nao embarcados nas imagens ainda)
