+++ src/backend/auth/tests/test_hello.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 
-from main import app
+from auth.main import app
 
 client = TestClient(app)
