+++ src/backend/colaboracao/fly.toml
@@ -1,8 +1,5 @@
-app = "movecity-backend"
+app = "movecity-colaboracao"
 primary_region = "gru"
-
-[build]
-  dockerfile = "Dockerfile"
 
 [http_service]
   internal_port = 8000
