+++ src/backend/gateway/requirements.txt
@@ -1,0 +1,4 @@
+fastapi>=0.115.0
+uvicorn[standard]>=0.30.0
+pydantic>=2.0
+pydantic-settings>=2.0
